requirements.txt
torch
transformers
pillow
streamlit
easyocr